requirements.txt
asgiref==3.11.0
Django
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.5
